% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{add_to_process}
\alias{add_to_process}
\title{Add process to a list of process}
\usage{
add_to_process(p_cmd, p_args = list(), verbose = T)
}
\arguments{
\item{p_cmd}{Command to be executed}

\item{p_args}{Arguments to be passed to the command}
}
\value{
Process ID from processx::new() which can be used for further monitoring of the process
}
\description{
This function adds process to a list and makes sure hte number of processes do not
exceed a user defined limit (< max_concurrent_jobs/cores/threads). Uses processx library to call
a child process
}
