% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{all2all_BLAST}
\alias{all2all_BLAST}
\title{Execute all2all BLAST}
\usage{
all2all_BLAST(
  first_list,
  second_list,
  blast_DB_dir = tempdir(),
  blast_program,
  output_dir = "./",
  blast_options = "",
  input_prefix_path = NULL,
  params_list = NULL,
  COMPLETE.format.ids = F,
  gzip.output = T,
  blast.sequence.limit = 10000,
  verbose = F
)
}
\arguments{
\item{first_list}{Vector of PATHS to FASTA}

\item{second_list}{Vector of PATHS to FASTA}

\item{blast_DB_dir}{Path to BLAST DBs, if provided, The Query and Subject FASTA are copied into this directory and then BLASTed}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program. Default is tempdir(). Can also be NULL}

\item{output_dir}{Path to BLAST output}

\item{blast_options}{Extra Options to be passed to the BLAST program}

\item{input_prefix_path}{If input lists/vectors are filenames, then provide input folder to prefix path}

\item{params_list}{Output of load_params() (Optional)}

\item{COMPLETE.format.ids}{Do BLAST Hit IDs of BLAST Hits (query and subject) have R-COMPLETE's long format IDs? (TRUE if using BLAST results from this package, FALSE (Default) otherwise) (Refer ?COMPLETE_PIPELINE_DESIGN) (Optional)}

\item{gzip.output}{Should the output files be Gzipped? Default - TRUE}

\item{blast.sequence.limit}{Maximum number of sequences to BLAST at a time. Default - 10000. If the query FASTA sequences > blast.sequence.limit, the sequences are split into multiple files and BLASTed}

\item{verbose}{Print DEBUG Messages?}
}
\value{
Named List of GRObjects (of BLAST Hits)
}
\description{
Executes All-to-All BLAST between two lists of organisms/genes/clusters. Output BLAST files are stored in the format filename1.filename2.all2all under output_dir. (All-to-All is simply Many-to-Many association)
}
\note{
ASSUMES Nucleotide sequences. Not checking/Not working for Protein/Peptide sequences.
}
\examples{
 all2all_BLAST(first_list = list.files(path="fasta",pattern = "\*.fa",all.files = T,full.names = T,include.dirs = F,recursive = F), second_list = list.files(path="fasta",pattern = "\*.fa",all.files = T,full.names = T,include.dirs = F,recursive = F),blast_DB_dir = "files/blastdb",blast_program = "tblastx",output_dir = "files/all2all",blast_options = "-strand plus",input_prefix_path = NULL,params_list=NULL)

}
