% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{calculate_gene_conservation}
\alias{calculate_gene_conservation}
\title{Internal Function - To calculate gene conservation score (GSC)}
\usage{
calculate_gene_conservation(
  blast_table,
  gene,
  score_col = 3,
  available_orgs,
  num_transcripts,
  REF_ORGS
)
}
\arguments{
\item{blast_table}{BLAST table}

\item{gene}{BLAST table}

\item{score_col}{Column name or number to be used as a score metric. Percentage Sequence Identity is used by default (column 3)}

\item{available_orgs}{Vector of available organisms (in the params_list$FASTA_OUT_PATH)}

\item{num_transcripts}{Number of Transcripts available for the gene. This can be obtained from the blastdb of a gene}

\item{REF_ORGS}{Vector of reference organisms}
}
\value{
Named Vector with an double value between minimum and maximum values in the score column (0-100 in case of percentage identity) which is the gene conservation score for the gene
}
\description{
This function is only designed to work with long ID format followed by the R-COMPLETE pipeline and assumes that the BLAST table is filtered by WISARD. The BLAST File/Table must be of the format 6. BLAST formats can be converted between each other using. By default, Percentage Sequence Identity is used as a score metric, the function expects BLAST Hits from tblastx (because the sequence is translated before it is BLASTed and the Percentage Identity is from the protein sequence). Other columns can be used as a metric with the score_col parameter
}
\details{
convert_BLAST_format(in_file,outfile = out_file,outformat=6,cols=c("qseqid","sseqid","pident","length","mismatch","gapopen","qstart","qend","sstart","send","evalue","bitscore","score","gaps","frames","qcovhsp","sstrand","qlen","slen","qseq","sseq","nident","positive"))
}
