% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{calculate_HSP_coverage}
\alias{calculate_HSP_coverage}
\title{Calculate HSP Coverage}
\usage{
calculate_HSP_coverage(
  blast_table,
  transcript_region_lengths,
  col.indices,
  min_coverage_filter = 0.5,
  COMPLETE.format.ids = F,
  params_list
)
}
\arguments{
\item{blast_table}{BLAST Table with Query->Subject Hits}

\item{transcript_region_lengths}{A data.frame (with one column for regional lengths and BLAST/Transcript IDs/Transcript Names as row.names) or a Named Vector or a Named List.}

\item{col.indices}{A Named List with indices of columns Query sequence ID (qseqid) and Subject sequence ID (sseqid). Eg col.indices=list(qseqid=1,sseqid=2)}

\item{min_coverage_filter}{Minimum HSP Coverage value to filter out Hits}

\item{COMPLETE.format.ids}{Do BLAST Hit IDs of BLAST Hits (query and subject) have R-COMPLETE's long format IDs? (TRUE if using BLAST results from this package, Default - FALSE otherwise) (Refer ?COMPLETE_PIPELINE_DESIGN)}

\item{params_list}{Output of load_params()}
}
\description{
This function calculates the coverage of HSPs (sum(HSP Alignment lengths)/mRNA CDS Length) between each Query->Subject Hits. Hits are filtered based on the minimum coverage filter value.
}
