% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{calculate_stats}
\alias{calculate_stats}
\title{Calculate Flank Values and other statistics from GTF Data}
\usage{
calculate_stats(gtf_data)
}
\arguments{
\item{gtf_data}{GTF data obtained from biomaRt for an organism}
}
\value{
Transcript Statistics from GTF data
}
\description{
This function calculates Flank Values (for transcripts whose UTR lengths are not known in which case Flank values are used instead of lengths to obtain the Flanking regions),
 CDS lengths, Number of Exon & CDS Blocks from GTF Data obtained from biomaRt. This function also checks for strandedness and only keeps the genes which are present in a strand (if requested by the user).
The GTF Data should have these columns c("chromosome_name", "strand", "ensembl_gene_id", "ensembl_transcript_id", "external_gene_name",
"start_position","end_position","exon_chrom_start","exon_chrom_end","transcript_version","transcript_length","5_utr_start","5_utr_end",
"transcript_start","transcription_start_site","transcript_end","cds_start","cds_end","3_utr_start","3_utr_end")
}
\details{
For calculating stats for actual GTF file, refer to extract_gtf_info.R in the github repo
}
