% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{check_files}
\alias{check_files}
\title{Checks FASTA files in a folder}
\usage{
check_files(fasta_path, org, genes, verbose = T, params_list)
}
\arguments{
\item{fasta_path}{(string) Path of folder to check}

\item{org}{(string) Name of the organism (format important, eg. "danio_rerio")}

\item{genes}{(vector/string) Filename/Vector of genes to check for}

\item{verbose}{(bool) Print check result messages?}

\item{params_list}{(string/list/COMPLETE-options) Output of load_params()}
}
\value{
(bool) TRUE if check passed, FALSE if check failed
}
\description{
This function checks if all the genes (and the transcript regions (cds,3utr,5utr)) are present
in the given folder path. The FASTA filenames should be of the format {gene}.{transcript region} for this
function to work.
}
\examples{
  check_files("files/fasta/notechis_scutatus",org = "notechis_scutatus",genes = "data/genelist.txt",verbose =T, params_list = load_params("pkg_data/parameters.txt"))

}
