% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{check_param}
\alias{check_param}
\title{Checks for parameter and return the value}
\usage{
check_param(
  param_table,
  param_id,
  optional = F,
  CAST_FUN = as.character,
  create_dir = F
)
}
\arguments{
\item{param_table}{parameter table or parameter file name}

\item{param_id}{ID of the paramter to be extracted from parameter file}

\item{gene_list}{File with a list of genes to extract data for(check the github repo for an example)}
}
\description{
If a parameter ID does not exist or if the value is empty and if the parameter is NOT optional then the execution is stopped
}
