% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-package.R
\docType{package}
\name{COMPLETE-package}
\alias{COMPLETE}
\alias{COMPLETE-package}
\title{COMPLETE: Comparative Orthologous Motif Prediction of Localization Elements with Transcript Extraction}
\description{
R package for extracting motifs of localization elements using a comparitive approach. 1) The package takes a list of genes and downloads protein coding transcripts of said genes for user provided organisms and organisms from Ensembl. 2) Groups the transcripts into ortholog/gene clusters and finds transcript level orthologs of genes/clusters with iterative BLAST approach (iBLAST). Ungrouped transcripts are optionally grouped during iBLAST. 3) Constructs Multiple Sequence Alignments (MSA) for clusters/genes. 4) Folds the MSAs and takes advantage of the evolutionary information to predict motifs which enable localization. Pipeline is split into multiple parts. The functions named in uppercase denote the functions which perform end-end execution of respective parts of the pipeline. Check ?COMPLETE_PIPELINE_DESIGN for more information.
}
\author{
\strong{Maintainer}: Vishvesh Karthik \email{vishveshkarthik@gmail.com}

}
\keyword{internal}
