% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{deindex_BLAST_table}
\alias{deindex_BLAST_table}
\title{De-Index BLAST Table Column}
\usage{
deindex_BLAST_table(blast_table, index_col)
}
\arguments{
\item{blast_table}{BLAST table}

\item{index_col}{Column index of BLAST table to de-index}
}
\value{
BLAST table with original IDs from index_col (index_col is removed from the table and replaced with original column)
}
\description{
Convert indexed FASTA IDs into original IDs. The indexed ID column is removed and replaced with original ID column
}
\note{
Indexed Column name must be prefixed with "indexed_" followed by original column name, eg "indexed_query"
}
\examples{
in1_data <- index_BLAST_table(in1_data,"sseqid",offset = 0)
in1_data <- index_BLAST_table(in1_data,"qseqid",offset = length(levels(factor( in1_data[,"qseqid"] ))))
#TO INDEX TWO TABLES
in1_data <- index_BLAST_table(in1_data,"sseqid",offset = 0)
in1_data <- index_BLAST_table(in1_data,"qseqid",offset = length(levels(factor( in2_data[,"qseqid"] ))) + length(levels(factor( in1_data[,"sseqid"] ))))
in2_data <- index_BLAST_table(in2_data,"sseqid",offset = length(levels(factor( in1_data[,"qseqid"] ))) + length(levels(factor( in2_data[,"sseqid"] ))))
in2_data <- index_BLAST_table(in2_data,"qseqid",offset = length(levels(factor( in2_data[,"qseqid"] ))) + length(levels(factor( in1_data[,"sseqid"] ))))

}
