% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{extract_transcript_orthologs}
\alias{extract_transcript_orthologs}
\title{Internal Function - Transcript Ortholog Extraction Function for R-COMPLETE pipeline}
\usage{
extract_transcript_orthologs(
  blast_program,
  blast_options,
  transcript_region = ".cds",
  run.mode = "both",
  params_list,
  clusters_left,
  clusters_right,
  input_dir,
  output_dir,
  clean_extract = F,
  verbose = F,
  seed = 123
)
}
\arguments{
\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program.}

\item{blast_options}{BLAST options for the BLAST program - COMPLETE::GetAvailableBLASTOptions()}

\item{transcript_region}{Transcript region (.cds, .3utr, .5utr, .exon, and other gtf regions(not tested)) (eg, ".cds" or ".fa"). Essentially this would be the file extensions of the input FASTA file(names)}

\item{run.mode}{"both" or "coverage_distance" (Default) or "coverage_filter" or "no_filter". "coverage_distance" - Hits are filtered based on distance between bi-directional minimum HSP coverages (coverage_distance <= min_coverage_filter). This option selects more BLAST hits and should be used when the coverage values are very low (and the BLAST Hits/sequences are distant). "coverage_filter" - Filters Hits based on minimum coverage of HSPs from either direction. Use this option when the coverage values are high (and the BLAST Hits/sequences are closely related). "both" - Uses both "coverage_distance" and "coverage_filter" and is very strict. "no_filter" - Only calculates HSP coverages and does not filter any Hits (Argument used for calculate_HSP_coverage())}

\item{params_list}{Output of load_params()}

\item{clusters_left}{Vector of file names (Set/Subset) in input_dir (OG Clusters/Genes) to BLAST clusters_right with (Can be same as clusters_right). This would be the FASTA file name(s) (without file extension)}

\item{clusters_right}{Vector of file names (Set/Subset) in input_dir (OG Clusters/Genes) to BLAST clusters_left with (Can be same as clusters_left). This would be the FASTA file name(s) (without file extension)}

\item{input_dir}{Give the directory with FASTA files (to BLAST between them using blast_program)}

\item{output_dir}{Directory for saving output files (\*.out, \*.all2all, \*.wis_out,\*.rbh_out)}

\item{clean_extract}{Delete Output file if exists? (Optional). Default - F}

\item{verbose}{Print DEBUG Messages?. Default - F}
}
\description{
This function calls the Transcript Ortholog Extraction pipeline which is used to reduce the pool of genes (step 1), reduce the pool of organisms and create sets of organisms (step 2), find transcript level orthologs (step 3). It takes only one argument which is the path/name of the BLAST program to use and refers to the values from the parameters file for other variables.
}
\details{
* Step 1 - Genes which are available in all the reference organisms are chosen
 * Step 2 - A Per-Gene Conservation Score (GSC) is calculated from the availability of a gene across organisms (literally the count of organisms which have the gene, normalized to 1 relative to other genes). Genes which have GSC score below GENE_DROP_THRESHOLD (parameter) are dropped (GENE_DROP_THRESHOLD=0 does not omit any genes). Sets of organisms are created based on the available genes after GSC filtering. I can suggest reference organisms based on which ones have the maximum number of genes
 * Step 3 - Two way BLAST followed by HSP selection with WISARD and Two way RBH are performed. Only transcripts which are bi-directional best hits are kept for further analysis (RBH from both the directions, not RBH in itself is bi-directionaly from the point of the QUERY, We can also do an RBH from the context of the SUBJECT to verify if it did not pas RBH by chance (even though it is very unlikely))
}
