% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA}
\alias{fetch_FASTA}
\title{Get FASTA data}
\usage{
fetch_FASTA(org_row)
}
\arguments{
\item{org_row}{Named vector with name of the organism (format important, eg. "danio_rerio") and other details eg, c(name="danio_rerio",version="106")}
}
\value{
Named Vector of the organism details
}
\description{
Main function to download FASTA data. This function wraps around other fetch_FASTA_*() functions (Except for fetch_FASTA_user() which calls fetch_FASTA() if a genome and a gtf are not provided)
}
