% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA_biomartr}
\alias{fetch_FASTA_biomartr}
\title{Get FASTA data from BIOMART (using biomartr)}
\usage{
fetch_FASTA_biomartr(org_row)
}
\arguments{
\item{org_row}{Named vector with name of the organism (format important, eg. "danio_rerio") and other details eg, c(name="danio_rerio",version="106")}
}
\value{
Named Vector of the organism details
}
\description{
This function downloads FASTA data from BIOMART using biomartr. The GENOME and GTF for the organism are downloaded and passed through the shell pipeline for extracting transcripts.
This function invokes external SHELL function extract_genomic_regions from system.file("exec", "functions.sh", mustWork = T ,package = "COMPLETE") (just like the piepline for user data) and cannot be monitored
}
