% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_stat}
\alias{fetch_stat}
\title{Load a Matrix}
\usage{
fetch_stat(x, col_name)
}
\arguments{
\item{x}{(data.frame) - Data Frame with columns named as required}

\item{col_name}{(string) - Name of the column for which the value is to be extracted}
}
\value{
A value which was requested from the GTF data frame
}
\description{
This function loads a file as a matrix. It assumes that the first column
contains the rownames and the subsequent columns are the sample identifiers.
Any rows with duplicated row names will be dropped with the first one being
kepted.
}
