% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{get_gtf_attributes}
\alias{get_gtf_attributes}
\title{Split GTF attributes into induvidual columns}
\usage{
get_gtf_attributes(gtf_data, attribute)
}
\arguments{
\item{gtf_data}{(data.frame) GTF Data}

\item{attribute}{(string) Attribute name}
}
\value{
(string) Attribute value
}
\description{
Split GTF attributes into induvidual columns
}
