% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{get_gtf_mart}
\alias{get_gtf_mart}
\title{Get GTF data from BIOMART (using biomaRt)}
\usage{
get_gtf_mart(org, gene_list)
}
\arguments{
\item{org}{Name of the organism (format important, eg. "danio_rerio")}

\item{gene_list}{Vector or File with genes to fetch the GTF data for}
}
\value{
GTF Data
}
\description{
This function downloads GTF data from BIOMART using biomaRt. The attributes downloaded are
c("chromosome_name", "strand", "ensembl_gene_id", "ensembl_transcript_id", "external_gene_name",
"start_position","end_position","exon_chrom_start","exon_chrom_end","transcript_version","transcript_length",
"5_utr_start","5_utr_end","transcript_start","transcription_start_site","transcript_end","cds_start","cds_end","3_utr_start","3_utr_end")
}
\examples{
    gtf_data <- get_gtf_mart(org, genes)

}
