% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{GetQuickBLASTEnums}
\alias{GetQuickBLASTEnums}
\title{Get an Instance of QuickBLASTEnums class and its exposed enumerations/properties}
\usage{
GetQuickBLASTEnums()
}
\value{
A QuickBLASTEnums Object
}
\description{
Get an Instance of QuickBLASTEnums class and its exposed enumerations/properties
}
