% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{GetQuickBLASTInstance}
\alias{GetQuickBLASTInstance}
\title{Get an Instance of QuickBLAST class and its exposed methods}
\usage{
GetQuickBLASTInstance(seq_type, strand, save_sequences)
}
\arguments{
\item{seq_type}{(integer) 0 for Nucleotide and 1 for Protein}

\item{strand}{(integer) 0 for Plus(+), 1 for Minus(-)}

\item{save_sequences}{(bool) Save Sequences in the arrow::RecordBatch (Only for BLASTing Files)}
}
\value{
A QuickBLAST Object
}
\description{
Get an Instance of QuickBLAST class and its exposed methods
}
