% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{GRObject_from_BLAST}
\alias{GRObject_from_BLAST}
\title{Create A GRanges Object from BLAST RESULTS}
\usage{
GRObject_from_BLAST(
  blast_input,
  col.names = NULL,
  COMPLETE.format.ids = F,
  col.indices = c(qseqid = 1, sseqid = 2, evalue = 11, qstart = 7, qend = 8, sstart = 9,
    send = 10, bitscore = 12, qcovhsp = 16, qlen = 18, slen = 19, frames = 15, pident =
    3, gaps = 14, length = 4, sstrand = 17)
)
}
\arguments{
\item{blast_input}{Filename with BLAST hits or a BLAST table}

\item{col.names}{Columns which for the BLAST tables/files}

\item{COMPLETE.format.ids}{Do BLAST Hit IDs of BLAST Hits (query and subject) have R-COMPLETE's long format IDs? (TRUE if using BLAST results from this package, FALSE otherwise)}

\item{col.indices}{A Named Vector with indices of columns Query sequence ID (qseqid), Subject sequence ID (sseqid), E Value (evalue), Query start (qstart), Query end (qend), Subject start (sstart), Subject end (send), Bitscore (bitscore), Query HSP coverage(qcovhsp), Query length (qlen), Subject length(slen), Frames (frames), Percentage Identity (pident), Number of Gaps (gaps), Alignment Length (length), Subject strand (sstrand). Default values are col.indices=c(qseqid=1,sseqid=2,evalue=11,qstart=7,qend=8,sstart=9,send=10,bitscore=12,qcovhsp=16,qlen=18,slen=19,frames=15,pident=3,gaps=14,length=4,sstrand=17)}
}
\value{
A GRanges object of BLAST hits
}
\description{
Functions accepts a filename or a BLAST table and converts it into a GenomicRanges::GRanges object (which can be used by R-WISARD). Give col.names when passing BLAST filenames. The accepted BLAST format is 6 and can be formatted from BLAST format 11 using blast_formatter (shown below). Columns indices can provided with the col.indices option. length(col.names) and ncol(blast_input) are assumed to be the same. BLAST table/file must be non-indexed (IDs not shortened) if COMPLETE.format.ids=T
}
\details{
blast_formatter -outfmt "6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore score gaps frames qcovhsp sstrand qlen slen qseq sseq nident positive"
}
