% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{group_FASTA}
\alias{group_FASTA}
\title{Group FASTA Sequences Based on a column index of COMPLETE.format.ids (?COMPLETE_PIPELINE_DESIGN)}
\usage{
group_FASTA(
  gene_list,
  params_list,
  id.col.index,
  remove.invalid.files = F,
  verbose = T
)
}
\arguments{
\item{gene_list}{Vector/Filename of Gene names (FASTA Filenames where the sequences are stored)}

\item{params_list}{Output of load_params()}

\item{id.col.index}{The index of Column of COMPLETE.format.ids (?COMPLETE_PIPELINE_DESIGN) to groups sequences into. Use id.col.index=1 for grouping sequences based on Transcript IDs, id.col.index=2 to group sequences into Organisms, id.col.index=3 for grouping sequences based on Gene Names and id.col.index=4 to groups sequences into Ortholog Clusters. Check COMPLETE$FORMAT_ID_INDEX for indices}

\item{remove.invalid.files}{Remove files which do not have COMPLETE.format.ids? Default - FALSE . (Refer ?COMPLETE_PIPELINE_DESIGN to know about COMPLETE.format.ids)}

\item{verbose}{Print DEBUG Messages?}
}
\description{
group_FASTA_genes() for run.more="gene" & group_FASTA_clusters() for run.mode="cluster" wrap this function. run.mode is used in FIND_TRANSCRIPT_ORTHOLOGS(). This function write groupings (clusters/genes) of each organism (org_name) into paste(params_list$OUT_PATH,"/genes/",org_name,"/ORG_CLUSTERS.", id.col.index,sep="")
}
