% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{group_FASTA_clusters}
\alias{group_FASTA_clusters}
\title{Group FASTA Sequences into Cluster Files}
\usage{
group_FASTA_clusters(params_list)
}
\arguments{
\item{params_list}{Output of load_params()}
}
\description{
Groups the Sequences in FASTA files in a folder(recursively) into Clusters and save the Cluster FASTA files in params_list$GROUPS_PATH. FASTA IDs are required to be in R-COMPLETE's long format (?COMPLETE_PIPELINE_DESIGN)
}
