% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{index_BLAST_tables}
\alias{index_BLAST_tables}
\title{Index BLAST Table(s)}
\usage{
index_BLAST_tables(
  blast_tables,
  query_index_cols,
  subject_index_cols,
  offset = 0
)
}
\arguments{
\item{blast_tables}{List of BLAST tables}

\item{query_index_cols}{List of Column indices of query sequence IDs from each table}

\item{subject_index_cols}{List of Column indices of subject sequence IDs from each table}

\item{offset}{Offset value to add to the indices}
}
\value{
BLAST table with indexed IDs from index_col (original index_col is attached to the table)
}
\description{
Convert longer FASTA IDs into short indices. BLAST table IDs need to be short for downstream processing (eg, graph package does not accept long names). The IDs are stored as an extra column and returned along with the BLAST table.
}
\note{
WARNING : Do not index it more than once (although you may do index it as much as you like)
}
\examples{
 in_data <- index_BLAST_tables(blast_tables = list(data1,data2),query_index_cols = 1, subject_index_cols = 2)

}
