% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{index_FASTA_IDs}
\alias{index_FASTA_IDs}
\title{Index FASTA IDs}
\usage{
index_FASTA_IDs(path, index_out)
}
\arguments{
\item{path}{Path with FASTA Files to index}

\item{index_out}{Output files with the indices of the format "file"[tab]"long_id"[tab]"index"}
}
\description{
Convert longer FASTA IDs into short indices. Indices are generated recursively for fasta fasta files within subfolders of the path provided.
}
