% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{label_seqIDs_lengthen}
\alias{label_seqIDs_lengthen}
\title{Label Sequence IDs of A FASTA File (Refer ?COMPLETE_PIPELINE_DESIGN about COMPLETE.format.IDs)}
\usage{
label_seqIDs_lengthen(
  fasta_path,
  org,
  db,
  org_ver,
  gene,
  odb_clusters,
  duplicates.method,
  params_list
)
}
\arguments{
\item{fasta_path}{(string) Path to FASTA File}

\item{org}{(string) Name of the organism}

\item{db}{(string) Query DB}

\item{org_ver}{(string) Version of the organism}

\item{gene}{(string/vector) Filename of gene list or Vector of gene names}

\item{odb_clusters}{(string) ODB Clusters (String collapsed with ',')}

\item{duplicates.method}{(string) merge/delete/make_unique. How to handle sequences with duplicate names?. For CDS/UTR blocks - merge (Concatenate sequences with duplicate names), for Exon blocks - make_unique (Sequence names/IDs are made unique), delete - for deleting all duplicate sequences (first seq is kept).}

\item{params_list}{(string/list/COMPLETE-options) Output of load_params()}
}
\description{
Convert short sequence IDs into longer COMPLETE format IDs
}
