% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{label_sequenceIDs}
\alias{label_sequenceIDs}
\title{Label Sequence IDs in FASTA Files}
\usage{
label_sequenceIDs(fasta_path, org, gene_list, odb_gene_map = NULL, params_list)
}
\arguments{
\item{fasta_path}{Path of FASTA Files}

\item{org}{Name of the organism}

\item{gene_list}{Filename of gene list or Vector of gene names}

\item{odb_gene_map}{Filename of OrthoDB Ortholog Clusters mapped to gene names  (Optional). If not provided, sequences cluster ID/name "ungrouped" is used}

\item{params_list}{Output of load_params()}
}
\description{
Convert short sequence IDs into longer COMPLETE format IDs
}
