% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{lengthen_FASTA_IDs}
\alias{lengthen_FASTA_IDs}
\title{Lengthen FASTA IDs}
\usage{
lengthen_FASTA_IDs(transcript_metadata, fasta_files = c())
}
\arguments{
\item{transcript_metadata}{Transcript Metadata file (from index_FASTA_IDs())}

\item{fasta_files}{Vector of FASTA files (Optional)}
}
\description{
WARNING - This will convert all the short-format(indices) into long-format FASTA IDs in the files.
}
