% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{LoadBLASTHits}
\alias{LoadBLASTHits}
\title{Load BLAST Hits into a data.frame}
\usage{
LoadBLASTHits(
  infile,
  transcript_ID_metadata = NULL,
  col.names = NULL,
  sep = "\\t",
  header = F,
  gzipped = F
)
}
\arguments{
\item{infile}{BLAST hits filename}

\item{transcript_ID_metadata}{Tab-delimited File with the filenames, indexed transcript IDs and the long transcript IDs.}

\item{col.names}{Name of the columns of the BLAST File}

\item{sep}{Delimiter of the BLAST File columns. Default - '\t'}

\item{header}{Does the file have a header? . Default - FALSE}

\item{gzipped}{Is the input file Gzipped? . Default - FALSE}
}
\value{
Data Frame with BLAST Results
}
\description{
Give the path to a BLAST Hits file to load it into a data.frame(BLAST HITs Table). The column names can be provided as col.names. Rows with NAs are automatically removed
}
\details{
Optional : You can provide the file/table with indexed Transcsript IDs. The format must be "file"[tab]"long_id"[tab]"index" (without a header). It can be generated with the  index_FASTA_IDs() (check ?index_FASTA_IDs or index_fastaIDs() in system.file("exec", "functions.sh", mustWork = T ,package = "COMPLETE"))
}
\note{
Column indices will be used when processing the data. Column names are only for user reference. First column must be the query sequence ID and the second column must be the subject sequence ID
}
\examples{
    LoadBLASTHits(infile, transcript_ID_metadata=NULL, col.names=NULL)

}
