% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-options.R
\name{mart_connect}
\alias{mart_connect}
\title{Connect to a biomaRt repo/dataset}
\usage{
mart_connect(MART_FUN = NULL, args = c(), verbose = T)
}
\arguments{
\item{MART_FUN}{Fuction to be called}

\item{args}{Named list of arguments to be used by the function}
}
\value{
Name of the dataset in biomaRt
}
\description{
This function tries to connect to BIOMART services (using biomaRt) and retries if the connection fails with a stepping timeout of 5secs.
The MART_FUN is called with do.call()
}
