% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{melt_wisard_list}
\alias{melt_wisard_list}
\title{Internal Function - Convert/Melt WISARD List to Data.Frame}
\usage{
melt_wisard_list(x)
}
\arguments{
\item{x}{WIZARD output List x}
}
\value{
Data Frame
}
\description{
Internal Function - Convert/Melt WISARD List to Data.Frame
}
