% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{merge_wisard_lists}
\alias{merge_wisard_lists}
\title{Internal Function - To Merge WISARD OUTPUTS Lists}
\usage{
merge_wisard_lists(x, y)
}
\arguments{
\item{x}{WIZARD output List x}

\item{y}{WIZARD output List y}
}
\value{
A Merged List of WISARD Outputs
}
\description{
Internal Function - To Merge WISARD OUTPUTS Lists
}
