% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{merge_wisard_table}
\alias{merge_wisard_table}
\title{Internal Function - To Merge WISARD OUTPUTS}
\usage{
merge_wisard_table(x, y)
}
\arguments{
\item{x}{WIZARD output x}

\item{y}{WIZARD output y}
}
\value{
A Merged WISARD Object
}
\description{
Internal Function - To Merge WISARD OUTPUTS
}
