% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{one2one_BLAST}
\alias{one2one_BLAST}
\title{Execute one2one BLAST}
\usage{
one2one_BLAST(
  first_list,
  second_list,
  file_ext = ".fa",
  run_name = "BLAST.one2one",
  blast_DB_dir = tempdir(),
  blast_program,
  output_dir = "./",
  blast_options = "",
  input_prefix_path = NULL,
  params_list = NULL,
  COMPLETE.format.ids = F,
  gzip.output = T,
  blast.sequence.limit = 0,
  clean_extract = F,
  n_threads = 8,
  verbose = F,
  seed = 123,
  return_data = TRUE
)
}
\arguments{
\item{first_list}{Vector of PATHS to FASTA}

\item{second_list}{Vector of PATHS to FASTA}

\item{file_ext}{File extension of input files. eg- ".cds" or ".fa"}

\item{run_name}{Name of the BLAST run and name of the output file}

\item{blast_DB_dir}{Path to BLAST DBs, if provided, The Query and Subject FASTA are copied into this directory and then BLASTed}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program. Default is tempdir(). Can also be NULL}

\item{output_dir}{Path to BLAST output}

\item{blast_options}{Extra Options to be passed to the BLAST program}

\item{input_prefix_path}{If input lists/vectors are filenames, then provide input folder to prefix path}

\item{params_list}{Output of load_params() (Optional)}

\item{COMPLETE.format.ids}{Do BLAST Hit IDs of BLAST Hits (query and subject) have R-COMPLETE's long format IDs? (TRUE if using BLAST results from this package, FALSE (Default) otherwise) (Refer ?COMPLETE_PIPELINE_DESIGN) (Optional)}

\item{gzip.output}{Should the output files be Gzipped? Default - TRUE}

\item{blast.sequence.limit}{Maximum number of sequences allowed in each BLAST file. Default - 0. If the query FASTA sequences > blast.sequence.limit, the sequences are split into multiple files and BLASTed. Use 0 to not split & copy the files into temporary files}

\item{clean_extract}{Delete Output file if exists? (Optional). Default - F}

\item{n_threads}{Number of Threads}

\item{verbose}{Print DEBUG Messages?}

\item{seed}{Seed value for reproducibility. Default - 123}

\item{return_data}{Logical. Should GRanges Object of BLAST Hits be returned? Default - TRUE}
}
\value{
A GRObject of BLAST Hits
}
\description{
Executes One-to-One BLAST between two lists of organisms/genes/clusters. The BLAST Hits are
}
\note{
ASSUMES Nucleotide sequences. Not checking/Not working for Protein/Peptide sequences.
}
\examples{
one2one_BLAST(first_set,second_set,blast_DB_dir=blast_DB_dir,blast_program,output_dir=output_dir, blast_options=blast_options, input_prefix_path=input_prefix_path, params_list=params_list,COMPLETE.format.ids=COMPLETE.format.ids, verbose=verbose)

}
