% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{print_toc}
\alias{print_toc}
\title{Pretty Print tictoc::toc() output}
\usage{
print_toc(clk)
}
\arguments{
\item{clk}{output of function tictoc::toc()}
}
\value{
(string) A formatted string to print to stdout/console
}
\description{
This function takes the output of tictoc::toc() and
pretty prints it to the console
}
