% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{RBH}
\alias{RBH}
\title{Find Reciprocal Blast Hits (RBH)}
\usage{
RBH(
  in1,
  in2,
  transcript_ID_metadata = NULL,
  col.names = NULL,
  index.tables = F,
  weight.col
)
}
\arguments{
\item{in1}{Input (query->subject) BLAST hits table/filename}

\item{in2}{Input (query<-subject) BLAST hits table/filename}

\item{transcript_ID_metadata}{Tab-delimited File with the filenames, indexed transcript IDs and the long transcript IDs.}

\item{col.names}{Columns which for the BLAST tables/files}

\item{index.tables}{Should the IDs in the tables be indexed?}

\item{weight.cols}{Column indices which must be used as edge weights (eg, mincovhsp - HSP coverage scores, bitscore etc)}
}
\value{
Named Vector of the organism details
}
\description{
Find RBH between BLAST results of different organisms/genes/transcripts (FASTA/FASTQ). The BLAST results must be of the format 6 and can be converted from BLAST format 11 with blast_formatter.
The command with the required column names are given below.
}
\details{
blast_formatter -outfmt "6 qseqid sseqid pident length mismatch gapopen qstart qend sstart send evalue bitscore score gaps frames qcovhsp sstrand qlen slen qseq sseq nident positive"

NOTE : Both the input files are expected to have the same number of columns with matching column order (and names)

Optional : You can provide the file/table with indexed Transcsript IDs. The format must be "file"[tab]"long_id"[tab]"index" (without a header). It can be generated with the  index_FASTA_IDs() (check ?index_FASTA_IDs or index_fastaIDs() in system.file("exec", "functions.sh", mustWork = T ,package = "COMPLETE"))
}
