% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{read.gtf}
\alias{read.gtf}
\title{Rgb - read.gtf implementation}
\usage{
read.gtf(
  file,
  attr = c("split", "intact", "skip"),
  features = NULL,
  quiet = FALSE
)
}
\arguments{
\item{file}{(string) Path to GTF}

\item{attr}{(string) Method to process GTF attributes. Accepted values are "split", "intact", "skip"}

\item{features}{(string) Vector with feature names to extract from GTF}

\item{quiet}{(bool) Print Messages? (Default: F)}
}
\value{
(data.frame) Extracted GTF
}
\description{
Sourcing and using read.gtf script from the deprecated R-Genome Browser (Rgb) package. Credits to the original author(s), Sylvain Mareschal <maressyl@gmail.com>
}
\author{
Sylvain Mareschal <maressyl@gmail.com>
}
