% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{run_QuickBLAST2Files}
\alias{run_QuickBLAST2Files}
\title{Perform Nucleotide BLAST between 2 FASTA files using the Custom Library}
\usage{
run_QuickBLAST2Files(
  query_path,
  subject_path,
  blast_out,
  blast.sequence.limit = 1000,
  n_threads = 8,
  blast_program,
  blast_options = ""
)
}
\arguments{
\item{query_path}{Path to Query FASTA File}

\item{subject_path}{Path to Subject FASTA File}

\item{blast_out}{Path to BLAST output file. Required for Files}

\item{blast.sequence.limit}{Maximum number of sequences to be BLASTed at a time}

\item{n_threads}{Number of threads. Default - 8}

\item{blast_program}{BLAST Program to use}

\item{blast_options}{BLAST Options to use}
}
\description{
BLAST between two FASTA files. BLAST output is saved in Arrow::Feather/Parquet format.
}
