% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{select_ref_org_groups}
\alias{select_ref_org_groups}
\title{Internal Function - Select groups which contain any of the reference organisms}
\usage{
select_ref_org_groups(params_list)
}
\arguments{
\item{params_list}{Output of load_params()}
}
\description{
This function removes groups from params_list$GROUPS_PATH which does not contain any/all of the reference organisms (based on params_list$SELECT_REF_ORG_GROUPS_METHOD)
}
