% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{shorten_FASTA_IDs}
\alias{shorten_FASTA_IDs}
\title{Shorten FASTA IDs}
\usage{
shorten_FASTA_IDs(transcript_metadata, fasta_files = c())
}
\arguments{
\item{transcript_metadata}{Transcript Metadata file (from index_FASTA_IDs())}

\item{fasta_files}{Vector of FASTA files to shorten IDs (Optional)}
}
\description{
WARNING - This will convert all the long-format FASTA IDs into short-format(indices) in the files for downstream compatibility. This can be reverted back with Lengthen FASTA IDs
}
