% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{transcript_ortholog_extraction}
\alias{transcript_ortholog_extraction}
\title{Transcript Ortholog Extraction Function for R-COMPLETE pipeline}
\usage{
transcript_ortholog_extraction(
  blast_program,
  blast_options = "",
  blast_DB_dir = NULL
)
}
\arguments{
\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program}

\item{blast_options}{Extra Options to be passed to the BLAST program}

\item{blast_DB_dir}{Path to BLAST DBs, if provided, The Query and Subject FASTA are copied into this directory and then BLASTed}
}
\description{
This function calls the Transcript Ortholog Extraction pipeline which is used to reduce the pool of genes (step 1), reduce the pool of organisms and create sets of organisms (step 2), find transcript level orthologs (step 3). It takes only one argument which is the path/name of the BLAST program to use and refers to the values from the parameters file for other variables.
}
\details{
* Step 1 - Genes which are available in all the reference organisms are chosen
 * Step 2 - A Per-Gene Conservation Score (GSC) is calculated from the availability of a gene across organisms (literally the count of organisms which have the gene, normalized to 1 relative to other genes). Genes which have GSC score below GENE_DROP_THRESHOLD (parameter) are dropped (GENE_DROP_THRESHOLD=0 does not omit any genes). Sets of organisms are created based on the available genes after GSC filtering. I can suggest reference organisms based on which ones have the maximum number of genes
 * Step 3 - Two way BLAST followed by HSP selection with WISARD and Two way RBH are performed. Only transcripts which are bi-directional best hits are kept for further analysis (RBH from both the directions, not RBH in itself is bi-directionaly from the point of the QUERY, We can also do an RBH from the context of the SUBJECT to verify if it did not pas RBH by chance (even though it is very unlikely))
}
