% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{all2all_BLAST}
\alias{all2all_BLAST}
\title{Execute all2all BLAST}
\usage{
all2all_BLAST(
  first_list,
  second_list,
  blast_DB_dir = NULL,
  blast_program,
  output_dir = "./",
  blast_options = "",
  input_prefix_path = NULL
)
}
\arguments{
\item{first_list}{Vector of PATHS to FASTA}

\item{second_list}{Vector of PATHS to FASTA}

\item{blast_DB_dir}{Path to BLAST DBs, if provided, The Query and Subject FASTA are copied into this directory and then BLASTed}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program}

\item{output_dir}{Path to BLAST output}

\item{blast_options}{Extra Options to be passed to the BLAST program}

\item{input_prefix_path}{If input lists/vectors are filenames, then provide input folder to prefix path}
}
\description{
Executes All-to-All BLAST between two lists of organisms/genes/clusters. Output BLAST files are stored in the format filename1.filename2.all2all under output_dir. ASSUMES Nucleotide sequences. Not checking/Not working for Protein/Peptide sequences.
}
