% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{check_files}
\alias{check_files}
\title{Checks FASTA files in a folder}
\usage{
check_files(fasta_path, org, genes, verbose = T, params_list)
}
\arguments{
\item{fasta_path}{Path of folder to check}

\item{org}{Name of the organism (format important, eg. "danio_rerio")}

\item{genes}{Vector of genes to check for}

\item{verbose}{Print check result messages?}

\item{params_list}{Output of load_params()}
}
\value{
TRUE if check passed, FALSE if check failed
}
\description{
This function checks if all the genes (and the transcript regions (cds,3utr,5utr)) are present
in the given folder path. The FASTA files should be of the format {gene}.{transcript region} for this
function to work.
}
\examples{
  check_files("files/fasta/notechis_scutatus",org = "notechis_scutatus",genes = "data/genelist.txt",verbose =T, params_list = load_params("data/parameters.txt"))

}
