% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{check_mart_dataset}
\alias{check_mart_dataset}
\title{Check if a dataset exists for the organism in BIOMART (through biomaRt)}
\usage{
check_mart_dataset(org)
}
\arguments{
\item{org}{Name of the organism}
}
\value{
Name of the dataset in biomaRt
}
\description{
This function accepts a name of an organism and
returns the name of the biomaRt dataset if available
}
