% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{COMPLETE_PIPELINE_DESIGN}
\alias{COMPLETE_PIPELINE_DESIGN}
\title{Design of R-COMPLETE}
\usage{
(1) EXTRACT_DATA()
(2) FIND_TRANSCRIPT_ORTHOLOGS()
}
\description{
Design of R-COMPLETE
}
\section{odb10v1_species.tab.gz - Ortho DB organism ids based on NCBI taxonomy ids (mostly species level) (https://v101.orthodb.org/download/odb10v1_species.tab.gz)}{

}

\section{odb10v1_genes.tab.gz  -Ortho DB genes with some info (https://v101.orthodb.org/download/odb10v1_genes.tab.gz)}{

}

\section{odb10v1_OG2genes.tab.gz - OGs to genes correspondence (https://v101.orthodb.org/download/odb10v1_OG2genes.tab.gz)}{
(OR)
}

\section{odb10v1_OGgenes_fixed.tab.gz - Merged & Transformed ODB file (Done within pipeline)}{

}

\section{odb10v1_OGgenes_fixed_user.tab.gz - Merged & Transformed ODB file BASED on user gene list (Done within pipeline)}{

}

\seealso{
\code{\link[=EXTRACT_DATA]{EXTRACT_DATA()}}, \code{\link[=FIND_ORTHOLOGS]{FIND_ORTHOLOGS()}}
}
\author{
Vishvesh Karthik (MDC-Berlin), \link{vishvesh.karthik@mdc-berlin.de}
}
