% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{COMPLETE_PIPELINE_DESIGN}
\alias{COMPLETE_PIPELINE_DESIGN}
\title{Design of R-COMPLETE}
\usage{
NULL

The pipeline uses R and BASH. BASH functions are invoked through R.
BASH functions are stored in system.file("exec", "functions.sh", mustWork = T ,package = "COMPLETE")

* REQUIRES/DEPENDANCIES :
     * Internet Connection
     * Referenced R packages
     * Linux with BASH ($SHELL must be set or /bin/bash must exist)
     * Parameters File (system.file("data", "parameters.txt", mustWork = T ,package = "COMPLETE"))
     * GNU parallel (in $PATH - BASH functions)
     * Samtools (in $PATH - BASH functions)
     * Bedtools (in $PATH - BASH functions)
     * OrthoDB (ODB) Flat Files (>= v10.1) (Pipeline is meant to work with v10.1) (https://www.orthodb.org/?page=filelist)
         # odb10v1_species.tab.gz - Ortho DB organism ids based on NCBI taxonomy ids (mostly species level) (https://v101.orthodb.org/download/odb10v1_species.tab.gz)
         # odb10v1_genes.tab.gz  -Ortho DB genes with some info (https://v101.orthodb.org/download/odb10v1_genes.tab.gz)
         # odb10v1_OG2genes.tab.gz - OGs to genes correspondence (https://v101.orthodb.org/download/odb10v1_OG2genes.tab.gz)
         (OR)
         # odb10v1_OGgenes_fixed.tab.gz - Merged & Transformed ODB file (Done within pipeline)
         # odb10v1_OGgenes_fixed_user.tab.gz - Merged & Transformed ODB file BASED on user gene list (Done within pipeline)

* PARAMETERS :
    The pipeline takes a single parameter file. This design was chosen
         1) To expose as many options as possible to the end-user.
         2) The pipeline uses BASH to perform some operations (which is significantly faster than R)
         and the parameter file is shared between R and BASH.
    The file is of the format [param_id==value==comment] where param_id and value columns are CASE-SENSITIVE
    (because its unnecessarily hard to check and convert param types in BASH). A default/example file is in
    system.file("data", "parameters.txt", mustWork = T ,package = "COMPLETE")

* FLOW :
    1) EXTRACT_DATA() - Extracts the transcript regions for Protein Coding Transcripts (provided in parameters, pipeline requires cds,5utr,3utr)
    from BIOMART and/or User provided genomes & GTFs. This functions uses biomaRt/biomartr for extracting data from BIOMART
    and BASH function extract_genomic_regions() for user provided data. Extraction priority/flow : User Data > biomaRt > biomartr
         * ODB Files are merged and transformed with BASH function merge_OG2genes_OrthoDB()
         * Orthologous genes are found for genes which are not present in the organism with BASH function check_OrthoDB()
         * Flank lengths are calculated from GTF data for missing UTRs (with variance correction, check ?calculate_gtf_stats)
         * FASTA Nucleotide Sequences for given TRANSCRIPT_REGIONS are fetched from BIOMART/Genome
         * Sequences are labelled with the following long ID format of R-COMPLETE (specific to this pipeline and referred to as COMPLETE.format.ids) (seqID_delimiter & transcripID_delimiter set in parameters, "::" & "||" respectively in this context )
              >$transcript_id $transcripID_delimiter $transcript_region ($strand) $seqID_delimiter $seqID_delimiter $org_name $gene_name $seqID_delimiter $ortho_cluster
              >SOME_TRANSCRIPT||cds(+)::SOMEORG::RANDOMGENE::ORTHOLOG_CLUSTERS
              >ENSDART00000193157||cds(+)::danio_rerio::sulf1::18335at7898,51668at7742,360590at33208
         * Sequences are grouped into files of ORTHOLOG_CLUSTERS

    2) FIND_ORTHOLOGS() -
}
\description{
Design of R-COMPLETE
}
\seealso{
\code{\link[=EXTRACT_DATA]{EXTRACT_DATA()}}, \code{\link[=FIND_ORTHOLOGS]{FIND_ORTHOLOGS()}}
}
\author{
Vishvesh Karthik (MDC-Berlin), \link{vishvesh.karthik@mdc-berlin.de}
}
