% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{convert_BLAST_format}
\alias{convert_BLAST_format}
\title{Convert between BLAST output formats}
\usage{
convert_BLAST_format(
  infile,
  outfile,
  outformat = 6,
  cols = c("qseqid", "sseqid", "pident", "length", "mismatch", "gapopen", "qstart",
    "qend", "sstart", "send", "evalue", "bitscore", "score", "gaps", "frames", "qcovhsp",
    "sstrand", "qlen", "slen", "qseq", "sseq", "nident", "positive"),
  conversion_prg = Sys.which("blast_formatter")
)
}
\arguments{
\item{infile}{Input BLAST Hits File}

\item{outfile}{Output BLAST Hits File}

\item{outformat}{Format to convert to (Default 6)}

\item{cols}{BLAST Columns to output. Check BLAST Output formats for more details}

\item{conversion_prg}{Path to blast_formatter if not found in $PATH}
}
\description{
Convert between formats using the blast_formatter program
}
\note{
blast_formatter must be present in $PATH and be accessible with Sys.which("blast_formatter"). (Optional) Input file must be of the BLAST format 11 to convert it between formats
}
\examples{
    convert_BLAST_format(in_file,outfile = out_file,outformat=6,cols=c("qseqid","sseqid","pident","length","mismatch","gapopen","qstart","qend","sstart","send","evalue","bitscore","score","gaps","frames","qcovhsp","sstrand","qlen","slen","qseq","sseq","nident","positive"))

}
