% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{EXTRACT_DATA}
\alias{EXTRACT_DATA}
\title{START HERE - Extracts Data}
\usage{
EXTRACT_DATA(param_file, gene_list)
}
\arguments{
\item{param_file}{Filename of a formatted parameter file (check the github repo for an example)}

\item{gene_list}{File with a list of genes to extract data for(check the github repo for an example)}
}
\description{
This is the main function which calls all the other functions and performs and end-end execution of data extraction part of the pipeline. It requires a filename of a formatted parameter file and a gene list (check the github repo for an example)
}
