% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{extract_transcript_regions}
\alias{extract_transcript_regions}
\title{Extract Transcripts from Genome}
\usage{
extract_transcript_regions(
  genome_path,
  gtf_path,
  gene_list,
  org_name,
  params_list,
  verbose
)
}
\arguments{
\item{genome_path}{Path to genom}

\item{gtf_path}{Path to annotation}

\item{gene_list}{Filename of gene list}

\item{org_name}{Name of the organism}

\item{params_list}{Parameter list from load_params()}
}
\description{
Requires a genome and an annotation. #' This function invokes external SHELL function extract_transcript_regions from fs::path_package("COMPLETE","exec","functions.sh") (just like the piepline for user data) and cannot be monitored
}
