% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA_biomartr}
\alias{fetch_FASTA_biomartr}
\title{Internal Function - Get FASTA data from BIOMART (using biomartr)}
\usage{
fetch_FASTA_biomartr(
  org_row,
  db = "",
  params_list,
  gene_list,
  data_types = c("gtf", "genome"),
  keep_data = F,
  ...
)
}
\arguments{
\item{org_row}{(Named vector) Named vector with name of the organism (format important, eg. "danio_rerio") and other details eg, c(name="danio_rerio",version="106",accession="acc1",taxid="dummy1")}

\item{db}{(string) Any supported DB from biomartr::listGenomes(), Valid Inputs: "ensembl", "genbank", "refseq", "" (Default: "ensembl"). Note: If empty, data is retrieved from both "genbank" and "ensembl"}

\item{params_list}{(string/list/COMPLETE-options) Output of load_params()}

\item{gene_list}{(string/vector) Vector or File containing list of genes}

\item{data_types}{(vector) Data types to fetch, one or all of "gtf", "genome" (Default: c("gtf","genome"))}

\item{keep_data}{(bool) Keep downloaded genomes and GTF data? (Default: FALSE)}

\item{only_fetch}{(bool) Only fetch data?, without extraction (only genomes & GTF annotations from biomartr) (Default: FALSE)}

\item{type}{(string) Filter to organisms. Input of biomartr::listGenomes() ["all", "kingdom", "group", "subgroup"] (Default: "all")}

\item{subset}{(string/vector) Filter to organisms. Input of biomartr::listGenomes() [Differs between DBs and filter type, check manually - e.g,if type == "group|subgroup" then check biomartr::getGroups()] (Default: NULL)}

\item{verbose}{(bool) Verbosity (Default: T)}
}
\value{
(Named vector) Organism details on successful extraction, NULL otherwise
}
\description{
This function downloads FASTA data from BIOMART using biomartr. The GENOME and GTF for the organism are downloaded and passed through the shell pipeline for extracting transcripts.
This function invokes external SHELL function extract_transcript_regions from fs::path_package("COMPLETE","exec","functions.sh") (just like the piepline for user data) and cannot be monitored
}
\examples{
    params_list <- load_params(fs::path_package("COMPLETE","pkg_data","parameters.txt"))
    fetch_FASTA_biomartr(c(name="danio_rerio",version="106",accession="acc1",taxid="dummy1"),params_list, gene_list = fs::path_package("COMPLETE","pkg_data","genelist.txt"))

}
