% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA_biomartr}
\alias{fetch_FASTA_biomartr}
\title{Internal Function - Get FASTA data from BIOMART (using biomartr)}
\usage{
fetch_FASTA_biomartr(
  org_row,
  params_list,
  gene_list,
  data_types = c("gtf", "genome"),
  only_fetch = F,
  keep_data = F,
  verbose = T
)
}
\arguments{
\item{org_row}{Named vector with name of the organism (format important, eg. "danio_rerio") and other details eg, c(name="danio_rerio",version="106")}

\item{params_list}{Output of load_params()}

\item{gene_list}{Vector or File containing list of genes}

\item{data_types}{Data types to fetch, one or all of "gtf", "genome" (Default: c("gtf","genome"))}

\item{only_fetch}{Only fetch data without extraction?}

\item{keep_data}{Keep downloaded genomes and GTF data?}

\item{verbose}{Print Messages?}
}
\value{
Named Vector of the organism details
}
\description{
This function downloads FASTA data from BIOMART using biomartr. The GENOME and GTF for the organism are downloaded and passed through the shell pipeline for extracting transcripts.
This function invokes external SHELL function extract_transcript_regions from fs::path_package("COMPLETE","exec","functions.sh") (just like the piepline for user data) and cannot be monitored
}
\examples{
    params_list <- load_params(fs::path_package("COMPLETE","pkg_data","parameters.txt"))
    fetch_FASTA_biomartr(c(name="danio_rerio",version="106"),params_list, gene_list = fs::path_package("COMPLETE","pkg_data","genelist.txt"))

}
