% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA_mart}
\alias{fetch_FASTA_mart}
\title{Get FASTA data from BIOMART (using biomaRt)}
\usage{
fetch_FASTA_mart(org, gtf_stats, fasta_path)
}
\arguments{
\item{org}{Name of the organism (format important, eg. "danio_rerio")}

\item{gtf_stats}{Data Frame from calculate_stats()}

\item{fasta_path}{Output path for transcript FASTA files}
}
\value{
Data Frame with transcripts for which FASTA was obtained
}
\description{
This function downloads FASTA data from BIOMART using biomaRt. The Data Frame from calculate_stats() is required.
}
