% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{fetch_FASTA_user}
\alias{fetch_FASTA_user}
\title{Get FASTA data}
\usage{
fetch_FASTA_user(data)
}
\arguments{
\item{data}{Named vector with name of the organism (format important, eg. "danio_rerio") and other details eg, c(name="danio_rerio",genome="http://some.link",gtf="some.file",version="106")}
}
\value{
Named Vector of the organism details
}
\description{
Function to download FASTA data for user data. fetch_FASTA() in invoked if a genome and a gtf are not provided.
}
