% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{FIND_TRANSCRIPT_ORTHOLOGS}
\alias{FIND_TRANSCRIPT_ORTHOLOGS}
\title{(2) - Find Transcript Orthologs}
\usage{
FIND_TRANSCRIPT_ORTHOLOGS(param_file, blast_program, gene_list)
}
\arguments{
\item{param_file}{Filename of a formatted parameter file (check the github repo for an example) or Output of load_params().}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program. BLAST options are taken from param_file}

\item{gene_list}{Vector or File with a list of genes to extract data for(check the github repo for an example).}
}
\description{
This function can be executed after COMPLETE::EXTRACT_DATA() and is the continuation of R-COMPLETE pipeline
}
\details{
This is the main function which calls all the other functions and performs and end-end execution of finding transcript level orthologs. It runs the iterative Transcript Ortholog Extraction pipeline which is used to reduce the pool of genes, reduce the pool of organisms, find transcript level orthologs (check ?extract_transcript_orthologs)
}
\note{
ONLY USE THIS FUNCTION WHEN RUNNING THE PIPELINE OF R-COMPLETE. Use other helper function to work with custom BLAST files not generated by this R package
}
