% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{FIND_TRANSCRIPT_ORTHOLOGS}
\alias{FIND_TRANSCRIPT_ORTHOLOGS}
\title{(2) - Find Transcript Orthologs}
\usage{
FIND_TRANSCRIPT_ORTHOLOGS(
  gene_list,
  params_list,
  blast_program = Sys.which("tblastx"),
  run.mode = COMPLETE$FORMAT_ID_INDEX$CLUSTERS
)
}
\arguments{
\item{gene_list}{Vector/Filename of gene names. (essentially Filenames of FASTA sequences)}

\item{params_list}{Filename of a formatted parameter file (check the github repo for an example) or Output of load_params().}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program. BLAST options are taken from params_list. Default is Sys.which("tblastx")}

\item{run.mode}{A value from COMPLETE$FORMAT_ID_INDEX. Default - COMPLETE$FORMAT_ID_INDEX$CLUSTERS. Find transcript orthologs in the level of Orgs, Genes or Ortholog Clusters. Genes have more tight orthology and fewer transcript orthologs which may be very similar. Ortholog Clusters are a level higher than Genes (Because an Ortholog Cluster can have more than one gene) and have highest number of transcript orthologs with a lot of dissimilarity. run.mode=COMPLETE$FORMAT_ID_INDEX$GENE is NOT RECOMMENDED because the sequences are grouped based on gene names, while run.mode=COMPLETE$FORMAT_ID_INDEX$CLUSTERS groups sequences based on protein identity.}
}
\description{
This function can be executed after COMPLETE::EXTRACT_DATA() and is the continuation of R-COMPLETE pipeline
}
\details{
This is the main function which calls all the other functions and performs and end-end execution of finding transcript level orthologs. It runs the iterative Transcript Ortholog Extraction pipeline which is used to reduce the pool of genes, reduce the pool of organisms, find transcript level orthologs (check ?extract_transcript_orthologs)
}
\note{
ONLY USE THIS FUNCTION WHEN RUNNING THE PIPELINE OF R-COMPLETE. Use other helper function to work with custom BLAST files not generated by this R package. run.mode="gene" is NOT RECOMMENDED because the sequences are grouped based on gene names
}
