% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{FIND_TRANSCRIPT_ORTHOLOGS}
\alias{FIND_TRANSCRIPT_ORTHOLOGS}
\title{START HERE - Find Transcript Orthologs}
\usage{
FIND_TRANSCRIPT_ORTHOLOGS(
  param_file,
  gene_list,
  blast_program = Sys.which("tblastx")
)
}
\arguments{
\item{param_file}{Filename of a formatted parameter file (check the github repo for an example)}

\item{gene_list}{File with a list of genes to extract data for(check the github repo for an example)}

\item{blast_program}{Give the name of the BLAST program to use (if in $PATH) or give the absolute path to the BLAST program (or Sys.which("tblastx"))}
}
\description{
This function can be executed after EXTRACT_DATA()
}
\details{
This is the main function which calls all the other functions and performs and end-end execution of finding transcript level orthologs. It runs the iterative Transcript Ortholog Extraction pipeline which is used to reduce the pool of genes, reduce the pool of organisms, find transcript level orthologs (check ?transcript_ortholog_extraction)

It requires a filename of a formatted parameter file and a gene list (check the github repo for an example).

ONLY USE THIS FUNCTION WHEN RUNNING THE PIPELINE OF COMPLETE. Use other helper function to work with custom BLAST files not generated by this R package
}
