% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{group_FASTA_clusters}
\alias{group_FASTA_clusters}
\title{Internal Function - Group FASTA Sequences into Cluster Files}
\usage{
group_FASTA_clusters(fasta_path)
}
\arguments{
\item{path}{Path with FASTA Files to index}
}
\description{
Groups the Sequences in FASTA files in a folder(recursively) into Clusters and save the Cluster FASTA files in complete_params$GROUPS_PATH. FASTA IDs are required to be in R-COMPLETE's long format
}
