% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{index_FASTA_IDs}
\alias{index_FASTA_IDs}
\title{Index FASTA IDs}
\usage{
index_FASTA_IDs(path = NULL, index_out = NULL)
}
\arguments{
\item{path}{Path with FASTA Files to index}

\item{index_out}{Output files with the indices of the format "file"\t"long_id"\t"index"}
}
\description{
Convert longer FASTA IDs into short indices. Indices are generated recursively. If path and index_out are not provided, the default FASTA_OUT_PATH and "files/rna_ids.txt" are used
}
