% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{make_BLAST_DB}
\alias{make_BLAST_DB}
\title{Make BLAST DB}
\usage{
make_BLAST_DB(fasta_file, clean_extract = F, verbose = F)
}
\arguments{
\item{fasta_file}{Path to FASTA File}

\item{clean_extract}{Remove existing BLAST DB?}

\item{verbose}{Print DEBUG Messages?}
}
\description{
Make BLAST DB
}
