% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-data_extraction.R
\name{read.gtf}
\alias{read.gtf}
\title{Rgb - read.gtf implementation}
\usage{
read.gtf(
  file,
  attr = c("split", "intact", "skip"),
  features = NULL,
  quiet = FALSE
)
}
\arguments{
\item{verbose}{Print Messages?}
}
\value{
Named Vector of the organism details
}
\description{
Sourcing and using read.gtf script from the deprecated R-Genome Browser (Rgb) package. Credits to the original author(s), Sylvain Mareschal <maressyl@gmail.com>
}
\author{
Sylvain Mareschal <maressyl@gmail.com>
}
