% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{run_BLAST}
\alias{run_BLAST}
\title{Perform Nucleotide BLAST}
\usage{
run_BLAST(
  query_path,
  subject_path,
  blast_DB_dir = NULL,
  blast_out,
  blast_program,
  run_name = "BLAST",
  blast_options = ""
)
}
\arguments{
\item{query_path}{Path to Query FASTA}

\item{subject_path}{Path to Subject FASTA}

\item{blast_DB_dir}{Path to BLAST DBs, if provided, The Query and Subject FASTA are copied into this directory and then BLASTed}

\item{blast_out}{Path to BLAST output file, Default BLAST FORMAT is 11}

\item{run_name}{Name of the BLAST run}

\item{blast_options}{Extra Options to be passed to the BLAST program}
}
\description{
BLAST between two organisms/genes/clusters. ASSUMES Nucleotide sequences. Not checking/Not working for Protein/Peptide sequences.
}
