% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{run_QuickBLAST}
\alias{run_QuickBLAST}
\title{Perform Nucleotide BLAST using the Custom Library}
\usage{
run_QuickBLAST(
  query,
  subject,
  blast_out,
  blast.sequence.limit,
  n_threads,
  blast_program,
  blast_options,
  input_type
)
}
\arguments{
\item{query}{Query FASTA (File or Seq)}

\item{subject}{Subject FASTA (File or Seq)}

\item{blast_out}{Path to BLAST output file. Required for Files, Not used for Seqs}

\item{blast.sequence.limit}{Maximum number of sequences to be BLASTed at a time, not used for Seqs}

\item{n_threads}{Number of threads. Default - 8}

\item{blast_program}{BLAST Program to use}

\item{blast_options}{BLAST Options to use}

\item{input_type}{(integer) 0 for Files, 1 for Sequences}
}
\description{
BLAST between two FASTA files or sequences. BLAST output is saved in Arrow::Feather/Parquet format (for files) or printed to stdout (for seqs).
}
