% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-BLAST_wrapper.R
\name{run_QuickBLAST2Seqs}
\alias{run_QuickBLAST2Seqs}
\title{Perform Nucleotide BLAST between 2 FASTA Seqs using the Custom Library}
\usage{
run_QuickBLAST2Seqs(query, subject, blast_program, blast_options = "")
}
\arguments{
\item{query}{Path to Query FASTA string}

\item{subject}{Path to Subject FASTA string}

\item{blast_program}{BLAST Program to use}

\item{blast_options}{BLAST Options to use}
}
\description{
BLAST between two FASTA Sequences. BLAST output is printed to stdout.
}
