% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COMPLETE-find_orthologs.R
\name{run_WISARD}
\alias{run_WISARD}
\title{Select highest scoring interval of non-overlapping HSPs from Bi-Directional BLAST Hits}
\usage{
run_WISARD(blast_hits, score_col, COMPLETE.format.ids = F)
}
\arguments{
\item{blast_hits}{GRanges Object of BLAST Hits (Query -> Subject)}

\item{score_col}{Column in the GRanges Object used for scoring intervals in WISARD}

\item{COMPLETE.format.ids}{Do BLAST Hit IDs of BLAST Hits (query and subject) have R-COMPLETE's long format IDs? (TRUE if using BLAST results from this package, FALSE otherwise)}
}
\value{
A GRanges object of BLAST hits
}
\description{
Select the highest scoring pairs (HSPs) which give the maximum coverage over the BLAST alignments of each transcript (without overlaps/minimal overlaps). These HSPs will then be used to find Transcript level orthologs across gene orthologs across organisms. This function only accepts bi-directional (Query <-> Subject) BLAST Hits formatted with GRObject_from_BLAST()
}
